Altera a estrutura de pastas dos repositories
