+++ grails-app/controllers/icaruswings/PayerController.groovy
@@ -1,7 +1,7 @@
 package icaruswings
 
 import icaruswings.adapters.PayerAdapter
-import icaruswings.utils.repositories.PayerRepository
+import icaruswings.repositories.PayerRepository
 
 class PayerController extends BaseController {
